requirements.txt
supabase
requests
google-generativeai
pytz
